pyproject.toml
[project]
name = "swrf"
version = "0.1.2"
description = "swrf / sorrowful - Check internet connection"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE.md"}
keywords = ["apachemq", "check",]
authors = [
  {name = "Cees van de Griend", email = "cees@griend.eu" }
]

maintainers = [
  {name = "Cees van de Griend", email = "cees@griend.eu" }
]

# Classifiers help users find your project by categorizing it.
#
# For a list of valid classifiers, see https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  # "Intended Audience :: Developers",
  # "Topic :: Software Development :: Build Tools",

  # Pick your license as you wish
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",

  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "requests",
  "stomp.py",
]

[project.urls]
"Homepage" = "https://github.com/ceesvandegriend/swrf"
"Source" = "https://github.com/ceesvandegriend/swrf"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"
